Move generate_experiment_project_name to server experiments package
